# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Media AI API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/style-images": {
      "get": {
        "tags": [
          "style-images"
        ],
        "summary": "GET style images",
        "operationId": "getStyleImages",
        "parameters": [
          {
            "name": "ipId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "GET tasks",
        "operationId": "getTasks",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoTaskList"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {}
        },
        "required": [
          "error"
        ],
        "description": "Standard error response."
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "description": "Standard success response."
      },
...
      "VideoTaskList": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "userId": {
              "type": "string"
            },
            "teamId": {
              "type": "string"
            },
            "workflowId": {
              "type": "string"
            },
            "ipId": {
              "type": [
                "string",
                "null"
              ]
            },
            "status": {
              "type": "string",
              "enum": [
                "PENDING",
                "RUNNING",
                "COMPLETED",
                "FAILED"
              ]
            },
            "params": {
              "type": [
                "string",
                "null"
              ]
            },
            "result": {
              "type": [
                "string",
                "null"
              ]
            },
            "error": {
              "type": [
                "string",
                "null"
              ]
            },
            "startedAt": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time",
              "description": "ISO 8601 date-time string."
            },
            "completedAt": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time",
              "description": "ISO 8601 date-time string."
            },
            "createdAt": {
              "type": "string",
              "format": "date-time",
              "description": "ISO 8601 date-time string."
            }
          },
          "required": [
            "id",
            "userId",
            "teamId",
            "workflowId",
            "ipId",
            "status",
            "params",
            "result",
            "error",
            "startedAt",
            "completedAt",
            "createdAt"
          ],
          "description": "Video generation task."
        }
      },
...
    },
...
  }
}